refactor(core): extract the OS input-injection layer into openlogi-inject
binding.rs was 3048 lines (3x the 1000-line limit) because it smuggled a three-platform event-synthesis engine — macOS CGEvent/NSEvent + dlopen'd private SPIs, Linux uinput + D-Bus, Windows SendInput — into a crate whose own docs promise it is platform- and IO-free. Move that whole effect layer into a new sibling crate, openlogi-inject, that depends on openlogi-core.

openlogi-core is now platform-free (no [target.*] deps), IO-free, and unsafe-free; binding.rs drops to 1382 lines. Action::execute* become free functions execute(action: &Action) in openlogi-inject (a downstream crate can't add inherent methods to core's Action); the public surface is execute / post_horizontal_scroll / SYNTHETIC_EVENT_USER_DATA / action_device_path — the four symbols the agent and hook consume. inject is published under the workspace's unified versioning, since openlogi-hook (published) now depends on it; release-plz.toml gains its entry. The moved SendInput block also gains the SAFETY comment it lacked.

Verified clean on macOS native and x86_64-pc-windows-gnu cross clippy (-D warnings), cargo doc (-D warnings), and the full core/inject/agent-core test suite. Linux cfg code is byte-identical to the original modulo the uniform import-path rewrite; it is checked by CI (no linux-gnu sysroot available locally).
