+++ crates/openlogi-core/Cargo.toml
@@ -23,23 +23,3 @@ tracing-subscriber = { workspace = true }
 
 [lints]
 workspace = true
-
-[target.'cfg(target_os = "linux")'.dependencies]
-evdev = "0.13"
-zbus = "5"
-
-[target.'cfg(target_os = "macos")'.dependencies]
-core-graphics = { version = "0.25.0", features = ["highsierra"] }
-core-foundation = { workspace = true }
-objc2 = "0.6.4"
-objc2-app-kit = { version = "0.3.2", features = [
-  "NSEvent",
-  "NSGraphicsContext",
-  "objc2-core-graphics",
-] }
-objc2-core-graphics = { version = "0.3.2", features = ["CGEvent"] }
-objc2-foundation = "0.3.2"
-
-# SendInput-based action synthesis (Action::execute_windows).
-[target.'cfg(target_os = "windows")'.dependencies]
-windows-sys = { workspace = true, features = ["Win32_UI_Input_KeyboardAndMouse"] }
