+++ crates/openlogi-hook/Cargo.toml
@@ -10,6 +10,7 @@ description = "OS-level mouse-event hook for OpenLogi. macOS via CGEventTap; Lin
 
 [dependencies]
 openlogi-core = { path = "../openlogi-core", version = "0.6.13" }
+openlogi-inject = { path = "../openlogi-inject", version = "0.6.13" }
 thiserror = { workspace = true }
 tracing = { workspace = true }
